Dedupe based on service name rather than EDPMServiceType
This change switches the behavior of the dedupe function to instead
deduplicate services based on their name rather than the EDPMServiceType. This allows
us to implement an enumerable list of possible EDPMServiceType values and ensure we
are always mounting the right directory for TLS compatibility.

Jira: https://issues.redhat.com/browse/OSPRH-10880
Signed-off-by: Brendan Shephard <bshephar@redhat.com>