Pass RELATED_IMAGES from controller-operator to controller-manager
This PR updates kustomize so that we have all RELATED_IMAGES
in the top level CSV, and thus they show up as relatedImages
there properly.

As the images now only get set on the controller-operator
they OSP service RELATED_IMAGES need to be passed to the
controller-manager so that they can be used accordingly.

To do this a go template loop is injected via the Makefile's
bindata target into the controller-managers deployment.
The loop is injected this way so as not to disturb the normal
kustomize config tree for the manager.

Jira: OSPRH-13075