Cleanup legacy OLM Subscription & CSVs
This cleans up legacy resources installed by OLM dependencies in
old releases by searching for specifically named resources.

This PR uses the unstructured client so that we don't have to import
structs (and dependencies).

NOTE: This PR does not yet address the leftover 'operator'
resources which persist and seem to get recreated even
if removed.