Reset deployment status when provisioned BMH hash changes
This would happen when a faulty node has been replaced by
deleting the BMH, openstack-baremetal-operator would provision
a new node from the available BMHs. We also need to tell the
user that they have to create a new deployment for the node
to be deployed.

Note: Cleaning up the deleted nodes from the services is still a
manual process.

jira: https://issues.redhat.com/browse/OSPRH-13948
Signed-off-by: rabi <ramishra@redhat.com>