[openstackclient] allow passing in custom environment variables
There is the need to be able to customize the environment variables
set in an openstackclient pod, e.g. to specify OpenStack API versions
to be used, `OS_COMPUTE_API_VERSION=2.95`. This change adds an
env list lot the CRD, like it is in a podSpect which can be used
to specify those, but also referencing them from e.g. a secret.

In an OpenStackControlPlane CR it would look like:
```
  openstackclient:
    template:
      env:
      - name: OS_IDENTITY_API_VERSION
        value: "3"
      - name: OS_COMPUTE_API_VERSION
        value: "2.95"
      - name: DB_PASSWORD
        valueFrom:
          secretKeyRef:
            key: password
            name: my-secret
```

Jira: https://issues.redhat.com/browse/OSPRH-13969

Signed-off-by: Martin Schuppert <mschuppert@redhat.com>