Add CA cert to bootstrap service
We want to add and trust the combined-ca-cert-bundle to the compute nodes to
ensure that any third party certs are trusted.  Lets use the bootstrap
service to do this.  This change will make sure that that bootstrap
service has the combined-ca-cert bundle is mounted to
/var/lib/openstack/cacerts/bootstrap.

We do this by making sure that any service which has caCerts defined
will mount that secret to /var/lib/openstack/cacerts/<service>

Jira: OSPRH-14205