Temporarily skip ListMustHaveSSATags
We recently introduced extraMounts in keystone, where we skip
listMustHaveSSATags. The reason behind that is related to the
fact that extraMounts, as per the crd-schema-checker, require
this tag, which is not defined in the most inner VolumeSource
type that we import in lib-common. To avoid addition work and
complication, and to not block any other patch in opentstack
operator, this patch skips the ListMustHaveSSATags validator
the same way we do in keystone-operator [1].

[1] https://github.com/openstack-k8s-operators/keystone-operator/blob/main/hack/crd-schema-checker.sh#L5C35-L5C56

Signed-off-by: Francesco Pantano <fpantano@redhat.com>