Common Notifications bus interface for services
Add a top level field in OpenStackControlPlaneSpec field that holds a
strusture with an enablement flag and a RabbitMQCluster CR name ref to
be used with notifications in the managed OpenStack instance.

Allow overriding it in OpenStack services templates, or disabling
notifications for all of them at the top level.

Provide a special handling for an empty value, which provides backwards
compatibility during oscp and services CRDs upgrades.

The proposed CRD design is expandable into supporting additional
messaging backends (provided via direct transport_url secret refs, or
the like).

Assume openstack-operator controller should implement a basic
inheritance mechanism from the top level field to the Nova service
(as a 1st implementation step example) based on the following rules:

if OpenStackControlPlaneSpec.NotificationsBus.Enabled then
  if novaSpecScore.NotificationBusInstance == nil then
    NovaSpec.NotificationsBusInstance := \
      OpenStackControlPlaneSpec.NotificationsBus.RabbitMQCluster
  else
    NovaSpec.NotificationsBusInstance := \
      NovaSpecScore.NotificationsBusInstance
else // not enabled
  NovaSpec.NotificationsBusInstance := ""

We leave further handling of empty values for the services specific
implementation.

Assume a similar pattern to be applied later for
KeystoneAPI.Spec.RabbitMqClusterName and other services that need to
consume or publish notifications to keep things consistent accross the
board..

Signed-off-by: Bohdan Dobrelia <bdobreli@redhat.com>