Common Notifications bus interface for services
Add a top level field in OpenStackControlPlaneSpec field that holds a
a RabbitMQ cluster CR name reference to be used for publishing and consuming
of notifications in the managed OpenStack instance.

Allow overriding it in OpenStack services templates.

Note about a special handling expected for an empty value by the services
that will be supporting this interface. It should provide backwards
compatibility during oscp and services CRDs upgrades.

There is no an empty value handling top scope (cannot disable
notifications top-scope as a cluster-wide), however. It may only take
a default value of a 'rabbitmq'. Use the service templates to override
it for an empty value, if needed.

The proposed CRD design is not expandable into supporting additional
messaging backends (provided via direct transport_url secret refs, or
the like). For that, the structure should be used instead of a simple string value.

Assume openstack-operator controller should implement a basic
inheritance mechanism from the top level field to the service template
based on the following rules:

if OSCP.Spec.<service_name>.Template.NotificationsBusInstance == nil
    OSCP.Spec.<service_name>.Template.NotificationsBusInstance = \
        OSCP.Spec.NotificationsBus.RabbitMqClusterName

We leave further handling of empty values for the services specific
implementation.

Assume a similar pattern to be applied later for
KeystoneAPI.Spec.RabbitMqClusterName and other services that need to
consume or publish notifications to keep things consistent across the
board.

Signed-off-by: Bohdan Dobrelia <bdobreli@redhat.com>