Add ServiceDefaults mechanism for updates support
ServiceDefaults is a new feild in the OpenStackVersion status
that can be used to manage service defaults during an update.

Older OpenStackVersions won't have newly initialized defaults set
and thus we treat them as legacy. In the case of glance
they would continue to deploy with httpd+proxypass until
the new OpenStackVersion for FR3 is selected and deployed.

Jira: https://issues.redhat.com/browse/OSPRH-16261