Bump default memory limit for controller manager to 512Mi
In tests it was seen that the openstack-operator controller manager
reaches just bellow the current set limit of 256Mi. This change
bumps the default to 512Mi for all controller managers, except
the openstack-operator-controller-operator which could be customized
using the CSV. It also introduces a MANAGER_MEMORY_LIMIT env var
in the CSV which can be used to change the memory limit of all
the other controller managers.

Jira: OSPRH-16802

Signed-off-by: Martin Schuppert <mschuppert@redhat.com>