Bump memory resources for openstack-operator-controller-manager
In tests it was seen that the openstack-operator controller
manager reaches just bellow the current set limit of 256Mi.
This change bumps the request to 512Mi and the limit to 1Gi
for the openstack-operator-controller-operator.

Jira: OSPRH-16802

Signed-off-by: Martin Schuppert <mschuppert@redhat.com>
(cherry picked from commit 9836480d5714900a88db0a1e679b552a18c9aa9b)