[OSPRH-16305] Introduce OpenStackLightspeed
Introduce the OpenStackLightspeed custom resource that manages the
deployment of Lightspeed-related functionalities. As of now, we expect
the OLS operator to be present in the cluster. Otherwise, enabling of
OpenStackLightspeed will fail. The Lightspeed functionalities can be
enabled through the OpenStackLightspeed as follows:

spec:
  llmEndpoint: <URL of the LLM>
  llmEndpointType: <endpoint type>
  llmCredentials: <secret containing token>
  modelName: <model name>
  tlsCACertBundle: <config map containg cert>

Only one OpenStackLightspeed CR can configure OpenStack Lightspeed at a
single time. If the user tries to manage the Lightspeed functionalities
through two or more OpenStackLightspeed instances, the user will be
warned, and the changes from the newest OpenStackLightspeed instance
will be ignored.