[OSPRH-16305] Introduce OpenStackLightspeed
Introduce the OpenStackLightspeed custom resource that manages the
deployment of Lightspeed-related functionalities. As of now, we expect
the OLS operator to be present in the cluster; otherise, enabling
OpenStackLightspeed will fail. The Lightspeed functionalities can be
enbled through OpenStackControlPlane:

openstacklightspeed:
  enabled: true
  template:
    llmEndpoint: "https://models.com"
    llmEndpointType: "rhelai_vllm"
    llmCredentials: <secret_name>
    modelName: "/data/granite-3.1-8b-instruct"
    TLSCACertBundle: <CM name>

Only one OpenStackControlPlane CR can configure OpenStack Lightspeed at
a single time. If the user tries to manage the Lightspeed
functionalities through two or more OpenStackControlPlanes, it will
fail.

Current limitations of the implementation:
 - We use uns.unstructured{} to work with OLSConfig [1] due to Go
   version incompatibilities between the OLS operator and
   openstack-operator.

 - The changes are cluster-wide. They will not work nicely if other
   services are installed.

[1] https://github.com/openshift/lightspeed-operator/blob/main/api/v1alpha1/olsconfig_types.go