Add AtLatestKnown condition
Add a condition to indicate if we are AtLatestKnown version.
Condition is set to `True` if `AvailableVersion` == `DeployedVersion`,
otherwize not set.

With this change we can do: `wait --for condition=MinorUpdateAvailable`,
then patch the targetVersion and wait for: `wait --for condition=AtLatestKnown`.

It is a nuance, but I think this can be useful for automation using oc wait.