Add RHOSO cluster_id to be set in the ansible variables
To integrate data insights data from the control plane (OCP)
collected by the insights-operator and data from the EDPM nodes, which
is collected by the insights-client we need to have some kind of RHOSO
cluster_id stored in the EDPM nodes somewhere.
This patch proposes to generate such RHOSO cluster_id as
<ocp_cluster_id>:<namespace> as such pair should clearly identify single
RHOSO cluster because there can not be more that one OpenStack cluster
deployed in single OCP namespace.
This ID is then set as ansible variable in the ansible group vars in the
generated inventory for each OpenStackDataPlaneNodeSet created and
edpm ansible role will then be able to write it on the edpm nodes.

Related: #OSPRH-19918
Signed-off-by: Slawek Kaplonski <skaplons@redhat.com>