Add CSV entries for new graphical console images
The GraphicalConsoles attribute is a feature flag and setting it to
Enabled won't be supported in the RHOSO-18 timeframe, this means
that there is not expected to be Antelope versions of the new images
IronicGraphicalConsoleImage and IronicNovncImage. These images will not
be pulled in the default case.

This change is proposed now to support future feature development.

Jira: OSPRH-20211