Fix potential panic
It's possible the cloudkitty enabled field will be nil. Especially
after an upgrade. Check the field before dereferencing and use
false if it's nil. Use the same check for Autoscaling and Ceilometer
enabled field just in case.

We might be able to simplify to what's described in the TODO comment,
but as noted there, the code wouldn't be 100% equivalent. And I'm
not sure if there was a reason to have the code like this. I also don't
want to risk destroying a working solution because of an edgecase I
didn't think of.