Improve Swift dataplane sample config readability
The Swift dataplane sample uses a slightly different intendation and
variable naming than the actual CR. This was still valid because "oc
kustomize" fixed this, eg. when using install_yamls.

However, if using this as a reference to configure your own CR this is
misleading. Fixing the variable name and intendation should improve
readability. Output of "oc kustomize" is exactly the same as before.

Signed-off-by: Christian Schwede <cschwede@redhat.com>