Add hack/go.work file as a static manifest of Go modules in the repo
go.work is the only way (native to Go packaging) to list what modules
make up a particular monorepo. It is very useful when programatically
extracting dependencies to prefetch prior to a hermetic build.

However it is also needed for developers to define their local
workspace, possibly including dependency injection and version overrides
that are never intended to be checked into the repository.

In order to satisfy both needs without forcing developers to change
their workflows, let's include a separate hack/go.work file for the
only purpose of being a manifest of all Go modules that are actually
part of the repository. This leaves the go.work file in the project's
root, never checked in, to be created and modified locally without
needing surgery every time changes are committed.

Also include a `gowork-manifest` make target that only needs to run when
we bump GOTOOLCHAIN_VERSION or add new modules to the project.