Don't Unmarshal edpm_services_override
edpm_services_override is already marshal'd to JSON in
ansible_execution.go, so there's no need to unmarshal and print it to
string, b/c go drops the commas from the marshal'd value.

This was breaking the download-cache service that relies on the proper
value of edpm_services_override when using download-cache and
servicesOverride on a Deployment.

Jira: OSPRH-21737
Signed-off-by: James Slagle <jslagle@redhat.com>