Marshal servicesOverride to JSON before formatting
go was formatting the servicesOverride slice without commas between the
items, which causes the value to be interpreted as a YAML list of one
string item. Marshal'ing the value to JSON first causes proper JSON
syntax so that the resulting YAML value is properly formatted with
commas.

This was breaking the download-cache service that relies on the proper
value of edpm_services_override when using download-cache and
servicesOverride on a Deployment.

Jira: OSPRH-21737
Signed-off-by: James Slagle <jslagle@redhat.com>